Moveall buildOptimisticSomething methods to ReportUtils
